requirements.txt
awscli==1.9.7
awsebcli==3.6.1
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
blessed==1.9.5
botocore==1.3.12
cement==2.4.0
certifi==2015.9.6.2
colorama==0.3.3
cycler==0.9.0
Django==1.8.6
django-allauth==0.24.1
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
email==4.0.2
httplib2==0.9.2
jmespath==0.9.0
lxml==3.5.0
matplotlib==1.5.0
mechanize==0.2.5
multiprocessing==2.6.2.1
numpy==1.10.1
oauthlib==1.0.3
pathspec==0.3.3
Pillow==3.0.0
pyasn1==0.1.9
PyAudio==0.2.9
pyparsing==2.0.6
python-dateutil==2.4.2
python-openid==2.2.5
pytz==2015.7
PyYAML==3.11
requests==2.6.2
requests-oauthlib==0.6.0
rsa==3.2.3
scipy==0.16.1
six==1.10.0
texttable==0.8.4
urllib3==1.12
virtualenv==13.1.2
Wave==0.0.2
wcwidth==0.1.5
websocket-client==0.34.0
